+++ geopandas/_vectorized.py
@@ -602,6 +602,18 @@ def boundary(data):
         return _unary_geo("boundary", data)
 
 
+def build_area(data):
+    if compat.USE_SHAPELY_20:
+        return shapely.build_area(data)
+    elif compat.USE_PYGEOS:
+        return pygeos.build_area(data)
+    else:
+        raise NotImplementedError(
+            f"shapely >= 2.0 or PyGEOS is required, "
+            f"version {shapely.__version__} is installed"
+        )
+
+
 def centroid(data):
     if compat.USE_SHAPELY_20:
         return shapely.centroid(data)